Use uint32_t for font face indices instead of size_t
As mentioned in [1], the number of fonts in a TrueType Collection is
always a uint32_t in the font format. As such, PDFium's font code should
also use uint32_t for face index values instead of size_t.

[1] https://pdfium-review.googlesource.com/c/pdfium/+/138932/comment/bfc8d03b_31c1d53f/

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/139010
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fxge/cfx_fontmapper.cpp
@@ -916,7 +916,7 @@ RetainPtr<CFX_Face> CFX_FontMapper::GetCachedTTCFace(void* font_handle,
                                                 std::move(font_data));
   }
   size_t font_offset = ttc_size - data_size;
-  size_t face_index =
+  uint32_t face_index =
       GetTTCIndex(font_desc->FontData().first(ttc_size), font_offset);
   RetainPtr<CFX_Face> face(font_desc->GetFace(face_index));
   if (face) {
